Fix segfault when using private buses (other than system and session).
When using private buses in Python (with python-e_dbus and python-dbus) the
program always segfaults trying to setup the integration of D-Bus with Ecore
mainloop. The problem is by the time the setup function is called the
connection with the private bus is not registered yet and then we're
segfaulting at this line of e_dbus_connection_new():

  cd->conn_name = strdup(dbus_bus_get_unique_name(conn));

This patch just adds a test to see if we're already registered or not. However,
we need to add proper support for private buses to e_dbus.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e17/libs/e_dbus@34115 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33