gustavo - i know you don't like this, but putting back old connman api support.
1. you broke connman legacy support during feature freeze. this isn't
a bug fix. it's a break.
2. newer connmans are broken for me - my netwoork basically barely
works. i can't make new connections reliably to anywhere outside my
lan, and even on my lan  ssh connectiosn justdrop all the time. i'm
sticking to a fdowngraded (0.55) connman exactly because of this.

i also noticed you broke api and abi for connman and ADDED a #define
now that peolpe have to agree that it is unstable - this isn't good in
managing a stable api. it's 1.0 - we maintain the connman api, and
it's our job to not break the api we expose. if connman upstream
breaks, it's our job to insulate and implement new features in a way
that maintains legacy support AND adds the new feature too. :)



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e_dbus@65406 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33